Change messages and chats id in SMessage to Long
Migrate types from Int to Long

+++ src/main/kotlin/com/csepanda/tgkot/data/SMessage.kt
@@ -6,7 +6,7 @@ package com.csepanda.tgbot.data
  *  @version 0.1
  *  @since   0.1
  */
-class SMessage private constructor(chatIdNum : Int?, chatIdStr : String?, text: String) {
+class SMessage private constructor(chatIdNum : Long?, chatIdStr : String?, text: String) {
     /** Unique identifier for the target chat or username of the target channel */
     val chatId: Any = chatIdNum?: chatIdStr!!
     /** Text of the message to be sent */
@@ -19,14 +19,14 @@ class SMessage private constructor(chatIdNum : Int?, chatIdStr : String?, text: 
     /** Sends the message silently. Users will receive a notification without sound */
     var disableNotification   : Boolean? = null; private set
     /** If the message is a reply, ID of the original message */
-    var replyToMessageId      : Int?     = null; private set
+    var replyToMessageId      : Long?    = null; private set
 
-    constructor(chatIdentifier: Int,    text: String) : this (chatIdentifier, null, text)
+    constructor(chatIdentifier: Long,   text: String) : this (chatIdentifier, null, text)
     constructor(toUserName    : String, text: String) : this (null, toUserName, text)
 
-    constructor(chatId              : Int,      text                  : String,
+    constructor(chatId              : Long,     text                  : String,
                 parseMode           : String?,  disableWebPagePreview : Boolean?,
-                disableNotification : Boolean?, replyToMessageId      : Int?)
+                disableNotification : Boolean?, replyToMessageId      : Long?)
             : this(chatId, text) {
         this.parseMode             = parseMode
         this.disableWebPagePreview = disableWebPagePreview
@@ -36,7 +36,7 @@ class SMessage private constructor(chatIdNum : Int?, chatIdStr : String?, text: 
 
     constructor(chatId              : String,   text                  : String,
                 parseMode           : String?,  disableWebPagePreview : Boolean?,
-                disableNotification : Boolean?, replyToMessageId      : Int?)
+                disableNotification : Boolean?, replyToMessageId      : Long?)
             : this(chatId, text) {
         this.parseMode             = parseMode
         this.disableWebPagePreview = disableWebPagePreview
